ardhi/requirements.txt
africastalking==1.2.4
celery==5.1.0
Django==3.2.3
django-colorful==1.3
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
django-geojson==3.2.0
django-heroku==0.3.1
django-leaflet==0.28.0
django-redis==4.12.1
django-tables2==2.4.0
djangorestframework==3.12.4
djangorestframework-gis==0.17
folium==0.12.1
geoip2==4.2.0
gunicorn==20.1.0
jsonfield==3.1.0
matplotlib==3.4.2
maxminddb==2.0.3
numpy==1.20.3
pandas==1.2.4
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.0
python-decouple==3.4
redis==3.5.3
reportlab==3.5.67
requests==2.25.1
segregation==1.5.0
simplejson==3.17.2
whitenoise==5.2.0

